Split model/ into a directory.
